Moved socket to module
